fix по ревью
