+++ packages/client/src/components/Forum/ForumMainListHeader/ForumMainListHeader.scss
@@ -1,0 +1,15 @@
+.forum-main-list-header {
+  display: grid;
+  grid-template-columns: 6fr 2fr 4fr;
+  line-height: 30px;
+  font-weight: bold;
+  margin-top: 10px;
+  background-color: rgba(255, 255, 255, .2);
+  text-align: left;
+
+  &__thread,
+  &__replies,
+  &__last-reply {
+    padding-left: 15px;
+  }
+}
