+++ packages/client/src/components/Forum/ForumPostList/ForumPostList.scss
@@ -1,0 +1,4 @@
+.forum-post-list {
+    list-style: none;
+    padding-left: 0;
+}
