+++ packages/client/src/components/Forum/ForumReply/ForumReply.scss
@@ -1,0 +1,58 @@
+.forum-reply {
+  display: grid;
+  grid-template-columns: 1fr 5fr;
+}
+
+.reply {
+  &__textarea {
+    width: 100%;
+    height: 130px;
+    border: 1px solid var(--color-lightgray);
+    text-align: start;
+    box-sizing: border-box;
+    background-color: var(--color-controls-off);
+    color: var(--color-controls-on);
+
+    &::placeholder {
+      font-style: italic;
+    }
+
+  }
+
+  &__input {
+    width: 60px;
+
+    &_hidden {
+      display: inline-block;
+      width: 0.1px;
+      height: 0.1px;
+      opacity: 0;
+      position: absolute;
+      z-index: -10;
+    }
+  }
+
+  &__label {
+    width: 30px;
+    height: 30px;
+    background-image: url('./../../../assets/paperclip-solid.png');
+    background-size: contain;
+    cursor: pointer;
+  }
+
+  &__button {
+    width: 200px;
+    height: 40px;
+
+    &_area {
+      display: flex;
+      flex-direction: row;
+      justify-content: flex-end;
+      align-items: flex-end;
+    }
+
+    &:disabled {
+      background-color: lightgrey;
+    }
+  }
+}
