+++ packages/client/src/components/Forum/ForumThreadList/ForumThreadList.scss
@@ -1,0 +1,4 @@
+.forum-thread-list {
+    list-style: none;
+    padding-left: 0;
+}
