+++ packages/client/src/components/Forum/ForumThreadList/ForumThreadList.tsx
@@ -1,0 +1,50 @@
+import React, { FC } from 'react';
+import { ForumThreadListItem } from './__Item/ForumThreadList__Item';
+import './ForumThreadList.scss';
+
+export interface topicsIncoming {
+  id: number;
+  closed: boolean;
+  title: string;
+  authorID: number;
+  lastReply?: number;
+  createdAt: string;
+  updatedAt: string;
+  User: {
+    displayName: string;
+    firstName: string;
+    secondName: string;
+    avatar: string;
+  };
+  Posts: {
+    message: string;
+    User: {
+      displayName: string;
+      firstName: string;
+      secondName: string;
+      avatar: string;
+    };
+    id: number;
+  }[];
+}
+
+export const ForumThreadList: FC<topicsIncoming[]> = (list: Record<number, topicsIncoming>) => {
+  const postList = Object.values(list);
+  return (
+    <ul className="forum-thread-list">
+      {postList.map((item, index) => (
+        <ForumThreadListItem
+          thread={item.title}
+          pageCount={item.Posts ? item.Posts.length / 10 : 1}
+          author={item.User.firstName}
+          startDate={item.createdAt.slice(0, 10)}
+          replies={item.Posts ? item.Posts.length.toString() : '0'}
+          lastReplyUser={item.User.firstName}
+          lastReplyDate={item.createdAt.slice(0, 10)}
+          threadID={item.id}
+          key={'thread' + index}
+        />
+      ))}
+    </ul>
+  );
+};
