+++ packages/client/src/components/Forum/ForumThreadList/__PageCounter/ForumThreadList__PageCounter.scss
@@ -1,0 +1,5 @@
+.forum-thread-list__page-counter {
+    color: var(--color-midgray);
+    margin: 0 3px;
+    cursor: pointer;
+  }
