+++ packages/client/src/components/Forum/ForumThreadPageContent/ForumThreadPageContent.scss
@@ -1,0 +1,17 @@
+
+.forum-thread-page-content {
+    padding-right: 20px;
+    color: var(--color-messages);
+    max-width: 1280px;
+
+    &__header {
+        font-size: 20px;
+        text-align: center;
+        line-height: 100px;
+    }
+
+    &__thread {
+        max-height: 426px;
+        overflow-y: auto;
+    }
+}
