--- packages/client/src/components/Phorum/PhorumPostList/PhorumPostList.scss
@@ -1,4 +1,0 @@
-.phorum-post-list {
-    list-style: none;
-    padding-left: 0;
-}
