--- packages/client/src/components/Phorum/PhorumThreadList/PhorumThreadList.scss
@@ -1,4 +1,0 @@
-.phorum-thread-list {
-    list-style: none;
-    padding-left: 0;
-}
