--- packages/client/src/components/Phorum/PhorumThreadList/__Item/PhorumThreadList__Item.scss
@@ -1,45 +1,0 @@
-.phorum-thread-list__item {
-  display: grid;
-  grid-template-columns: 6fr 2fr 4fr;
-  border-bottom: 1px solid var(--color-midgray);
-}
-
-.thread {
-  text-align: left;
-  padding: 15px;
-
-  &__info {
-    color: var(--color-midgray);
-    margin-top: 10px;
-  }
-
-  &__text {
-    cursor: pointer;
-  }
-}
-
-.replies {
-  text-align: left;
-  padding: 15px;
-}
-
-.last-reply {
-  text-align: left;
-  padding: 15px;
-
-  &__user {
-    text-decoration: underline;
-    cursor: pointer;
-  
-    &_theme_dark {
-      color: var(--color-text-dark_theme);
-    }
-  }
-
-  &__date {
-    color: var(--color-midgray);
-    text-decoration: underline;
-    margin-top: 10px;
-    cursor: pointer;
-  }
-}
