--- packages/client/src/components/Phorum/PhorumThreadList/__PageCounter/PhorumThreadList__PageCounter.scss
@@ -1,5 +1,0 @@
-.phorum-thread-list__page-counter {
-    color: var(--color-midgray);
-    margin: 0 3px;
-    cursor: pointer;
-  }
