--- packages/client/src/components/Phorum/PhorumThreadPageContent/PhorumThreadPageContent.scss
@@ -1,17 +1,0 @@
-
-.phorum-thread-page-content {
-    padding-right: 20px;
-    color: var(--color-messages);
-    max-width: 1280px;
-
-    &__header {
-        font-size: 20px;
-        text-align: center;
-        line-height: 100px;
-    }
-
-    &__thread {
-        max-height: 426px;
-        overflow-y: auto;
-    }
-}
