+++ packages/client/src/pages/Forum/ForumMainPage/ForumMainPage.scss
@@ -1,0 +1,3 @@
+.forum-wrapper {
+  text-align: center;
+}
