+++ packages/client/src/pages/Forum/ForumThreadPage/ForumThreadPage.scss
@@ -1,0 +1,3 @@
+.forum-wrapper {
+  text-align: center;
+}
