--- packages/client/src/pages/Phorum/PhorumMainPage/PhorumMainPage.scss
@@ -1,3 +1,0 @@
-.phorum-wrapper {
-  text-align: center;
-}
