--- packages/client/src/pages/Phorum/PhorumThreadPage/PhorumThreadPage.scss
@@ -1,3 +1,0 @@
-.phorum-wrapper {
-  text-align: center;
-}
