--- packages/server/app/utils/sendRes.ts
@@ -1,12 +1,0 @@
-import type { Response } from "express";
-
-export async function query<T>(fn: () => Promise<T>, res: Response, error?: string) {
-    return  fn().then(data => {
-        res.send(JSON.stringify(data));
-      })
-      .catch(err => {
-        res.status(500).send({
-          message: err.message || error,
-        })
-      })
-  };
